requirements.txt
pytest~=7.1.2
cryptography~=37.0.2
requests~=2.27.1
python-dotenv~=0.20.0
PyAutoGUI~=0.9.53